feat: Updated user Interface and cleaned up scripts
