+++ src/unity/Virtual Museum/Assets/Scripts/DisplayTime.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using TMPro;
+using UnityEngine;
+
+public class DisplayTime : MonoBehaviour
+{
+    public TMP_Text timedisplay;
+
+    // Update is called once per frame
+    void Update()
+    {
+        timedisplay.text = "Zeit: " + StandardFlag.currentTime.ToString();
+    }
+}
